requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.121.3
h11==0.16.0
idna==3.11
pillow==12.0.0
pydantic==2.12.4
pydantic_core==2.41.5
python-multipart==0.0.20
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
